Reorganize again, update doc, redirected references
