+++ src/util/roll.js
